LPS-97899 Social User Statistics - Apply new SF rule: Classes that are not exported should be in internal package
