+++ modules/apps/deprecated/social-user-statistics-web/src/main/java/com/liferay/social/user/statistics/web/internal/configuration/definition/SocialUserStatisticsPortletInstanceConfigurationBeanDeclaration.java
@@ -12,29 +12,23 @@
  * details.
  */
 
-package com.liferay.social.user.statistics.web.configuration.definition;
+package com.liferay.social.user.statistics.web.internal.configuration.definition;
 
-import com.liferay.portal.kernel.settings.definition.ConfigurationPidMapping;
-import com.liferay.social.user.statistics.constants.SocialUserStatisticsPortletKeys;
-import com.liferay.social.user.statistics.web.configuration.SocialUserStatisticsPortletInstanceConfiguration;
+import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
+import com.liferay.social.user.statistics.web.internal.configuration.SocialUserStatisticsPortletInstanceConfiguration;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Kailash Yadav
+ * @author Iván Zaera
  */
-@Component(service = ConfigurationPidMapping.class)
-public class SocialUserStatisticsPortletInstanceConfigurationPidMapping
-	implements ConfigurationPidMapping {
+@Component(service = ConfigurationBeanDeclaration.class)
+public class SocialUserStatisticsPortletInstanceConfigurationBeanDeclaration
+	implements ConfigurationBeanDeclaration {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
 		return SocialUserStatisticsPortletInstanceConfiguration.class;
-	}
-
-	@Override
-	public String getConfigurationPid() {
-		return SocialUserStatisticsPortletKeys.SOCIAL_USER_STATISTICS;
 	}
 
 }
