+++ modules/apps/deprecated/social-user-statistics-web/src/main/resources/META-INF/resources/init.jsp
@@ -46,7 +46,7 @@ page import="com.liferay.social.kernel.service.SocialActivityCounterLocalService
 page import="com.liferay.social.kernel.util.SocialConfigurationUtil" %><%@
 page import="com.liferay.social.kernel.util.comparator.SocialActivityCounterNameComparator" %><%@
 page import="com.liferay.social.user.statistics.constants.SocialUserStatisticsPortletKeys" %><%@
-page import="com.liferay.social.user.statistics.web.configuration.SocialUserStatisticsPortletInstanceConfiguration" %><%@
+page import="com.liferay.social.user.statistics.web.internal.configuration.SocialUserStatisticsPortletInstanceConfiguration" %><%@
 page import="com.liferay.taglib.search.ResultRow" %>
 
 <%@ page import="java.util.ArrayList" %><%@
